Get neko working on heroku again
Mongolab upgraded the database to 3.x so I've upgraded the pymongo
library and updated neko.py and webneko.py to use the syntax of the new
version.

One of the dependencies of the bleach library was choking on heroku and
I couldn't figure out how to fix it (not a python dev) so I removed the
couple lines of code using it.